dme/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
kafka-python>=2.0.2
reportlab>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
